crypto-key-pair-schnorr
